refactor: handling 404 cases
