+++ src/components/CompanyAndJobTitle/InterviewExperiences/index.js
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import CompanyAndJobTitleWrapper from '../CompanyAndJobTitleWrapper';
-import StatusRenderer from '../StatusRenderer';
+import { BoxStatusRenderer } from '../StatusRenderer';
 import InterviewExperiencesSection from './InterviewExperiences';
 import InterviewExperienceHelmet from './Helmet';
 import Searchbar from '../Searchbar';
@@ -11,7 +11,6 @@ const InterviewExperiences = ({
   pageName,
   tabType,
   interviewExperiences,
-  status,
   page,
   pageSize,
   totalCount,
@@ -22,25 +21,34 @@ const InterviewExperiences = ({
     pageName={pageName}
     tabType={tabType}
   >
-    <InterviewExperienceHelmet
+    <Searchbar pageType={pageType} tabType={tabType} />
+    <BoxStatusRenderer
       pageType={pageType}
       pageName={pageName}
-      interviewExperiences={interviewExperiences}
-      page={page}
+      tabType={tabType}
+      render={() => {
+        return (
+          <Fragment>
+            <InterviewExperienceHelmet
+              pageType={pageType}
+              pageName={pageName}
+              interviewExperiences={interviewExperiences}
+              page={page}
+            />
+            <InterviewExperiencesSection
+              pageType={pageType}
+              pageName={pageName}
+              tabType={tabType}
+              data={interviewExperiences}
+              page={page}
+              pageSize={pageSize}
+              totalCount={totalCount}
+              canView={canView}
+            />
+          </Fragment>
+        );
+      }}
     />
-    <Searchbar pageType={pageType} tabType={tabType} />
-    <StatusRenderer status={status}>
-      <InterviewExperiencesSection
-        pageType={pageType}
-        pageName={pageName}
-        tabType={tabType}
-        data={interviewExperiences}
-        page={page}
-        pageSize={pageSize}
-        totalCount={totalCount}
-        canView={canView}
-      />
-    </StatusRenderer>
   </CompanyAndJobTitleWrapper>
 );
 
@@ -51,7 +59,6 @@ InterviewExperiences.propTypes = {
   pageName: PropTypes.string.isRequired,
   pageSize: PropTypes.number.isRequired,
   pageType: PropTypes.string.isRequired,
-  status: PropTypes.string.isRequired,
   tabType: PropTypes.string.isRequired,
   totalCount: PropTypes.number.isRequired,
 };
