+++ src/components/CompanyAndJobTitle/TimeAndSalary/index.js
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 
 import CompanyAndJobTitleWrapper from '../CompanyAndJobTitleWrapper';
-import StatusRenderer from '../StatusRenderer';
+import { BoxStatusRenderer } from '../StatusRenderer';
 import TimeAndSalarySection from './TimeAndSalary';
 import Helmet from './Helmet';
 import OvertimeSection from './OvertimeSection';
@@ -14,7 +14,6 @@ const TimeAndSalary = ({
   tabType,
   salaryWorkTimes,
   salaryWorkTimeStatistics,
-  status,
   page,
   pageSize,
   totalCount,
@@ -24,27 +23,36 @@ const TimeAndSalary = ({
     pageName={pageName}
     tabType={tabType}
   >
-    <Helmet
-      pageType={pageType}
-      pageName={pageName}
-      salaryWorkTimes={salaryWorkTimes}
-      salaryWorkTimeStatistics={salaryWorkTimeStatistics}
-      page={page}
-    />
     <OvertimeSection statistics={salaryWorkTimeStatistics} />
     <Searchbar pageType={pageType} tabType={tabType} />
-    <StatusRenderer status={status}>
-      <TimeAndSalarySection
-        pageType={pageType}
-        pageName={pageName}
-        tabType={tabType}
-        salaryWorkTimes={salaryWorkTimes}
-        salaryWorkTimeStatistics={salaryWorkTimeStatistics}
-        page={page}
-        pageSize={pageSize}
-        totalCount={totalCount}
-      />
-    </StatusRenderer>
+    <BoxStatusRenderer
+      pageType={pageType}
+      pageName={pageName}
+      tabType={tabType}
+      render={() => {
+        return (
+          <Fragment>
+            <Helmet
+              pageType={pageType}
+              pageName={pageName}
+              salaryWorkTimes={salaryWorkTimes}
+              salaryWorkTimeStatistics={salaryWorkTimeStatistics}
+              page={page}
+            />
+            <TimeAndSalarySection
+              pageType={pageType}
+              pageName={pageName}
+              tabType={tabType}
+              salaryWorkTimes={salaryWorkTimes}
+              salaryWorkTimeStatistics={salaryWorkTimeStatistics}
+              page={page}
+              pageSize={pageSize}
+              totalCount={totalCount}
+            />
+          </Fragment>
+        );
+      }}
+    />
   </CompanyAndJobTitleWrapper>
 );
 
@@ -55,7 +63,6 @@ TimeAndSalary.propTypes = {
   pageType: PropTypes.string,
   salaryWorkTimeStatistics: PropTypes.object.isRequired,
   salaryWorkTimes: PropTypes.array,
-  status: PropTypes.string.isRequired,
   tabType: PropTypes.string,
   totalCount: PropTypes.number.isRequired,
 };
