+++ src/components/CompanyAndJobTitle/WorkExperiences/index.js
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import CompanyAndJobTitleWrapper from '../CompanyAndJobTitleWrapper';
-import StatusRenderer from '../StatusRenderer';
+import { BoxStatusRenderer } from '../StatusRenderer';
 import WorkExperiencesSection from './WorkExperiences';
 import Helmet from './Helmet';
 import Searchbar from '../Searchbar';
@@ -11,7 +11,6 @@ const WorkExperiences = ({
   pageName,
   tabType,
   workExperiences,
-  status,
   page,
   pageSize,
   totalCount,
@@ -22,25 +21,34 @@ const WorkExperiences = ({
     pageName={pageName}
     tabType={tabType}
   >
-    <Helmet
+    <Searchbar pageType={pageType} tabType={tabType} />
+    <BoxStatusRenderer
       pageType={pageType}
       pageName={pageName}
-      workExperiences={workExperiences}
-      page={page}
+      tabType={tabType}
+      render={() => {
+        return (
+          <Fragment>
+            <Helmet
+              pageType={pageType}
+              pageName={pageName}
+              workExperiences={workExperiences}
+              page={page}
+            />
+            <WorkExperiencesSection
+              pageType={pageType}
+              pageName={pageName}
+              tabType={tabType}
+              data={workExperiences}
+              page={page}
+              pageSize={pageSize}
+              totalCount={totalCount}
+              canView={canView}
+            />
+          </Fragment>
+        );
+      }}
     />
-    <Searchbar pageType={pageType} tabType={tabType} />
-    <StatusRenderer status={status}>
-      <WorkExperiencesSection
-        pageType={pageType}
-        pageName={pageName}
-        tabType={tabType}
-        data={workExperiences}
-        page={page}
-        pageSize={pageSize}
-        totalCount={totalCount}
-        canView={canView}
-      />
-    </StatusRenderer>
   </CompanyAndJobTitleWrapper>
 );
 
@@ -50,7 +58,6 @@ WorkExperiences.propTypes = {
   pageName: PropTypes.string.isRequired,
   pageSize: PropTypes.number.isRequired,
   pageType: PropTypes.string.isRequired,
-  status: PropTypes.string.isRequired,
   tabType: PropTypes.string.isRequired,
   totalCount: PropTypes.number.isRequired,
   workExperiences: PropTypes.arrayOf(PropTypes.object),
